% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardExport.R
\name{dashboardExport}
\alias{dashboardExport}
\title{dashboardExport}
\usage{
dashboardExport(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  achillesDatabaseSchema = resultsDatabaseSchema,
  smallCellCount = 5,
  outputFolder = "output",
  databaseId = NULL,
  cdmVersion = NULL,
  verboseMode = TRUE
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function
\code{createConnectionDetails} or \code{createDbiConnectionDetails} in
the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Fully qualified name of database schema that contains OMOP CDM schema.
On SQL Server, this should specifiy both the database and the schema,
so for example, on SQL Server, 'cdm_instance.dbo'.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write results to.
On SQL Server, this should specifiy both the database and the schema,
so for example, on SQL Server, 'cdm_results.dbo'.}

\item{achillesDatabaseSchema}{(OPTIONAL) Fully qualified name of database schema where the Achilles results
tables can be found (achilles_results, achilles_results_dist).
On SQL Server, this should specifiy both the database and the schema,
so for example, on SQL Server, 'cdm_results.dbo'.}

\item{smallCellCount}{(OPTIONAL) To avoid patient identifiability, cells with small counts
(<= smallCellCount) are deleted. Set to NULL if you don't want any deletions.
Default = 5.}

\item{outputFolder}{(OPTIONAL) Path to write the export and store logs and SQL files.
If the path does not exist, it is created. Defaults to 'output'.}

\item{databaseId}{(OPTIONAL) Name of the source, used in the filename exported. If not given, it is retrieved from cdm_source table.}

\item{cdmVersion}{(OPTIONAL) Version number of the OMOP CDM. If not given, it is retrieved from cdm_source table.}

\item{verboseMode}{(OPTIONAL) Boolean to determine if the console will show all execution steps. Default = TRUE}
}
\value{
dataframe of exported results (invisibly)
}
\description{
\code{dashboardExport} exports a set of descriptive statistics summary from the CDM,
to be uploaded in the Database Dashboard.
}
\details{
\code{dashboardExport} exports the results from Achilles, stored in the achilles_results
and achilles_results_dist tables, to a single csv file.
This csv file can be uploaded to the Database Dashboard entry in the DARWIN-EU(R) Portal.
There are two measures to prevent sharing of too detailed information:
\enumerate{
\item all counts are rounded up to the nearest hundred.
\item a small cell count is applied and counts smaller than this record are omitted (default=5)
This is a light-weight version of the EHDEN CatalogueExport, where the Achilles analyses were rerun.
}
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms="sql server", server="your_server")
dashboardExport(
     connectionDetails = connectionDetails,
     cdmDatabaseSchema = "cdm",
     resultsDatabaseSchema = "results",
     outputFolder = "output",
     databaseId = "MyName"
)
}
}
