% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardExport.R
\name{getAnalysisIdsToExport}
\alias{getAnalysisIdsToExport}
\title{Get Achilles analysis ids to be exported}
\usage{
getAnalysisIdsToExport()
}
\value{
vector of integer analysis ids
}
\description{
Get Achilles analysis ids to be exported
}
