% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashboardExport.R
\name{getRequiredAnalysisIds}
\alias{getRequiredAnalysisIds}
\title{Get minimally required Achilles analysis ids, used in the DARWIN Database Dashboard}
\usage{
getRequiredAnalysisIds()
}
\value{
vector of integer analysis ids
}
\description{
Get minimally required Achilles analysis ids, used in the DARWIN Database Dashboard
}
