% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResults.R
\name{exportResults}
\alias{exportResults}
\title{exportAnalyses}
\usage{
exportResults(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema,
  outputFolder,
  databaseId,
  achillesDatabaseSchema = resultsDatabaseSchema,
  smallCellCount = 5,
  analysisIds = getAnalysisIdsToExport()
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function
\code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{Fully qualified name of database schema that contains OMOP CDM schema.
On SQL Server, this should specifiy both the database and the schema,
so for example, on SQL Server, 'cdm_instance.dbo'.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write results to.
On SQL Server, this should specifiy both the database and the schema,
so for example, on SQL Server, 'cdm_results.dbo'.}

\item{outputFolder}{Path to store logs and SQL files}

\item{databaseId}{Name of the source, used in the filename exported}

\item{achillesDatabaseSchema}{(OPTIONAL) Fully qualified name of database schema where the Achilles results
tables can be found (achilles_results, achilles_results_dist).
On SQL Server, this should specifiy both the database and the schema,
so for example, on SQL Server, 'cdm_results.dbo'.}

\item{smallCellCount}{To avoid patient identifiability, cells with small counts
(<= smallCellCount) are deleted. Set to NULL if you don't want any deletions.
Default = 5.}

\item{analysisIds}{List of analysis ids to export. Default is to export all analyses.}
}
\description{
Export Achilles and DashboardExport analysis results to a single csv file.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms="sql server", server="your_server")
.executeDEAnalyses(
  connectionDetails = connectionDetails,
  cdmDatabaseSchema = cdmDatabaseSchema,
  resultsDatabaseSchema = resultsDatabaseSchema
)
exportResults(
  connectionDetails = connectionDetails,
  cdmDatabaseSchema = cdmDatabaseSchema,
  resultsDatabaseSchema = resultsDatabaseSchema,
  achillesDatabaseSchema = achillesDatabaseSchema,
  smallCellCount = smallCellCount,
  analysisIds = analysisIds,
  outputFolder = outputFolder,
  databaseId = databaseId
)
}
}
